Move get_default_primary_set_objective to automl/utils and clean up docstrings.
