+++ evalml/automl/__init__.py
@@ -1,2 +1,3 @@
-from .automl_search import AutoMLSearch, get_default_primary_search_objective
+from .automl_search import AutoMLSearch
+from .utils import get_default_primary_search_objective
 from .data_splitters import TrainingValidationSplit
