+++ evalml/automl/automl_search.py
@@ -17,6 +17,7 @@ from .pipeline_search_plots import PipelineSearchPlots
 
 from evalml.automl.automl_algorithm import IterativeAlgorithm
 from evalml.automl.data_splitters import TrainingValidationSplit
+from evalml.automl.utils import get_default_primary_search_objective
 from evalml.data_checks import (
     AutoMLDataChecks,
     DataChecks,
@@ -54,22 +55,6 @@ from evalml.utils.logger import (
 )
 
 logger = get_logger(__file__)
-
-
-def get_default_primary_search_objective(problem_type):
-    """Get the default primary search objective for a problem type.
-
-    Arguments:
-        problem_type (str or ProblemType): problem type of interest.
-
-    Returns:
-        ObjectiveBase: primary objective instance for the problem type.
-    """
-    problem_type = handle_problem_types(problem_type)
-    objective_name = {'binary': 'Log Loss Binary',
-                      'multiclass': 'Log Loss Multiclass',
-                      'regression': 'R2'}[problem_type.value]
-    return get_objective(objective_name, return_instance=True)
 
 
 class AutoMLSearch:
